added parser utility
This is a tool to help a program written in C# to gather and collect log files and put them all in one place. The script then sends the information to the program. This feature wasn't in the original program so this auxilery script is used in lieu of a feature in the origianl program.